Optimize ripgrep filter with early exit on empty files, support fractional timeouts in CLI and config default values

+++ context_builder/cli.py
@@ -611,7 +611,7 @@ def main():
     parser.add_argument("--max-interface-depth", type=int, default=None)
     parser.add_argument("--disable-pruning", action="store_true", default=None)
     parser.add_argument("--lsp-timeout", type=int, default=None)
-    parser.add_argument("--ripgrep-timeout", type=int, default=None)
+    parser.add_argument("--ripgrep-timeout", type=float, default=None)
     parser.add_argument("--no-language-server", action="store_true", default=None)
     parser.add_argument("--skip-ffi", action="store_true", default=None)
     parser.add_argument("--skip-macro-expansion", action="store_true", default=None)
